Make getURLs a reusable function.
